+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Changed
+- Added `Request.header/3` method 
 - Renamed client_name to user_agent_suffix
 - Use package name as prefix instead of ExApiClient.
 - set Mix.Project.config[:app] as a default value for user_agent_suffix